[#514] Move all network dependencies to data layer: OkHttp + Chucker
